+++ app/Presenters/TransactionPresenter.php
@@ -2,6 +2,7 @@
 
 namespace App\Presenters;
 
+use Illuminate\Support\Carbon;
 use PlacetoPay\ThreeDsSecureBase\Constants\Versions\V220\TransactionStatus;
 
 class TransactionPresenter extends Presenter
@@ -21,5 +22,16 @@ class TransactionPresenter extends Presenter
     public function deviceChannel(): string
     {
         return trans("fields.device_channel.{$this->entity->device_channel}");
+    }
+
+    public function locationData(): array|null
+    {
+        $data = $this->entity->getIpInformation();
+
+        $data['created_at'] = Carbon::parse(data_get($data, 'created_at', Carbon::now()))
+            ->setTimezone(auth()->user()->timezone ?? config('app.timezone'))
+            ->toDateTimeString();
+
+        return $data;
     }
 }